Graph: Added accessors to get the out degree of a vertex
Because we are following a lazy deletion strategy, in the query
we want to check whether the out degree of a candidate neighbor
is zero, if yes, it was marked as deleted before, we should
remove it from our candidate list based on which we will return
nearest neighbors results.

get_out_degree will return the number of outgoing edges from the
vertex.