Graph: Added deletion markers
Added a new unordered_set indices_deleted to know which indices have
already been deleted, and modified the tostring() to only write back
those vertices that were not deleted, and those edges without deleted
source or end vertex. In this way, we will follow the lazy deletion
strategy and also give us a chance to completely remove those deleted
nodes and affected edges when writing the graph back to disk.