Graph: modified and deleted unused functions
Modified add_vertex_if_new() to not only check if it is already
in name_to_index but also check whether it was previously
deleted, this is to avoid problem when you insert a vertex which
was previously lazy deleted. Because it was not removed from
name_to_index but just labeled as deleted in indices_deleted.

Modified remove_vertex() which will label the node in indices_deleted
as deleted, and also will remove all the outgoing edges from it.

Also deleted all the functions that was not used.