[WIP] Modified LSHEuclideanHashFunction
Following the implementation of E2LSH, the parameters we need for
Euclidean Hash Function is a random vector generated from Guassian
distribution and a random uniform number generated from w which should
be provided by user as the seed.

When writing back to a file, the random uniform number b and seed w will
be appended to the vector values.

The final hash value is not finished yet. It will combine all the hash
values and generate a final integer.