Adding queue of hash functions
The operations of mysql is asynchronous, so the oldest
hash function returned by the query might be supposed
deleted, but that deletion is not finished yet due to
the asynchronous nature of mysql. The solution of locking
tables might be harmful to performance, so here we add a queue
of hash functions to solve this problem.